python-docker-k8s-workshop/week 1/use-case/server/requirements.txt
anyio==3.5.0
asgiref==3.5.0
click==8.1.2
distlib==0.3.4
dnspython==2.2.1
fastapi==0.75.0
fastapi-router-controller==0.5.0
filelock==3.6.0
h11==0.13.0
idna==3.3
platformdirs==2.5.1
pydantic==1.9.0
pymongo==4.0.2
six==1.16.0
sniffio==1.2.0
starlette==0.17.1
typing_extensions==4.1.1
uvicorn[standard]
virtualenv==20.14.0
python-dotenv